feat: Cache pool and descriptors in kernel, register kernels

+++ include/enum_iterator.hpp
@@ -21,7 +21,7 @@ namespace rllm
         {}
 
         // Number of inner (v2) values per outer (v1) step.
-        size_t inner_size() const noexcept { return m_end - m_start; }
+        size_t inner_size() const noexcept { return static_cast<size_t>(m_end) - static_cast<size_t>(m_start); }
 
         Enum operator*() const
         {
